+++ src/models/PreviewOrderResponseOneOfOrderInputOneOf1.ts
@@ -34,6 +34,13 @@ import {
   PreviewOrderResponseOneOfOrderInputOneOfPrescriptionsInnerToJSON,
   PreviewOrderResponseOneOfOrderInputOneOfPrescriptionsInnerToJSONTyped,
 } from "./PreviewOrderResponseOneOfOrderInputOneOfPrescriptionsInner";
+import type { CreateOrderRequestPrescriber } from "./CreateOrderRequestPrescriber";
+import {
+  CreateOrderRequestPrescriberFromJSON,
+  CreateOrderRequestPrescriberFromJSONTyped,
+  CreateOrderRequestPrescriberToJSON,
+  CreateOrderRequestPrescriberToJSONTyped,
+} from "./CreateOrderRequestPrescriber";
 
 /**
  *
@@ -59,6 +66,12 @@ export interface PreviewOrderResponseOneOfOrderInputOneOf1 {
    * @memberof PreviewOrderResponseOneOfOrderInputOneOf1
    */
   userId?: string | null;
+  /**
+   *
+   * @type {CreateOrderRequestPrescriber}
+   * @memberof PreviewOrderResponseOneOfOrderInputOneOf1
+   */
+  prescriber?: CreateOrderRequestPrescriber | null;
   /**
    *
    * @type {string}
@@ -117,6 +130,10 @@ export function PreviewOrderResponseOneOfOrderInputOneOf1FromJSONTyped(
         : (json["otcItems"] as Array<any>).map(CreateOrderRequestOtcItemsInnerFromJSON),
     practiceId: json["practiceId"],
     userId: json["userId"] == null ? undefined : json["userId"],
+    prescriber:
+      json["prescriber"] == null
+        ? undefined
+        : CreateOrderRequestPrescriberFromJSON(json["prescriber"]),
     shippingAddressId: json["shippingAddressId"] == null ? undefined : json["shippingAddressId"],
     externalOrderId: json["externalOrderId"] == null ? undefined : json["externalOrderId"],
     prescriptions: (json["prescriptions"] as Array<any>).map(
@@ -147,6 +164,7 @@ export function PreviewOrderResponseOneOfOrderInputOneOf1ToJSONTyped(
         : (value["otcItems"] as Array<any>).map(CreateOrderRequestOtcItemsInnerToJSON),
     practiceId: value["practiceId"],
     userId: value["userId"],
+    prescriber: CreateOrderRequestPrescriberToJSON(value["prescriber"]),
     shippingAddressId: value["shippingAddressId"],
     externalOrderId: value["externalOrderId"],
     prescriptions: (value["prescriptions"] as Array<any>).map(